// src/user/interfaces/user-response.interface.ts
import { IAuditFields } from "src/common/interfaces/common.interface";
import { IUserBaseProfile } from "./user-base-response.interface";

export interface IUserResponse extends IUserBaseProfile, IAuditFields {
    isEmailVerified: boolean;
    isEmailPublic: boolean;
    role: string;
    status: string;
    lastLoginAt: Date;
}

export interface IUserResponseWithRecoveryCodes extends IUserBaseProfile {
    recoveryCodes: string[];
    createdAt: Date;
}